feat: add obs script for fast updating text
To reduce the CPU load and make it easier to display in OBS, we also add
a divider so that we can update every N frames, rather than ever frame.
We also limit the display to the latest 4 lines.

This limit is important because there can often be extremely long
utterances when someone is reading a talk.