refactor: separate en-US behavior and use '`' to switch
The models for en-US are significantly faster/better than es-MX. So,
when translating from en-US to other languages, use the enhanced models
and how realtime results. But, when going from others to en-US, only
show completed chunks. This seems to result in the best
quality/experience.

Switching the toggle to a rarely used char like '`' allows the operator
to type normally without accidentally triggering a toggle.